This uses https://github.com/hendrikmuhs/ccache-action to set up https://ccache.dev/ together with https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/caching-dependencies-to-speed-up-workflows to create cache entries at https://github.com/iree-org/iree/actions/caches and then pull them down at the start of each build job.

* Right now the cache keys are just `${{ github.job }}-${{ matrix.name }}`, which will create entries named like `ccache-build_test_runtime-ubuntu-20.04-2024-08-01T22:19:46.787Z`.
  * We can turn off the timestamps with `append-timestamp: false` if we find that our 10GB cache limit gets overrun enough to have workflows get cache misses (PkgCI already produces 800MB cache entries, so we're always floating way over that limit and old entries continually get evicted).
  * We could do something more sophisticated like put the LLVM commit hash in the cache entry name, but that shouldn't matter for runtime builds 🤞
* The Windows build in particular has large binaries (mainly HAL CTS tests). Reducing the size there will speed up link time during building and hopefully also trim the cache size.
* I'm matching the `write-caches` logic we have in other workflows for now, which should maintain cache integrity (only reviewed + merged code writes to the cache) while also limiting the number of entries we generate.

## Sample job run

https://github.com/ScottTodd/iree/actions/runs/10206588680/job/28239799198

* Load from cache:
  ![image](https://github.com/user-attachments/assets/62fca6d6-c45a-4efd-8628-558096dcf643)

* Save to cache:
  ![image](https://github.com/user-attachments/assets/bfac1b30-2c3c-40a4-89b1-05b334a5d4f9)

* View cache files (https://github.com/ScottTodd/iree/actions/caches?query=key%3Accache):
  ![image](https://github.com/user-attachments/assets/a51bc82b-6778-41fc-9d5d-76df0beace39)

---

Also
* dropped support for using Docker containers in the build/test runtime jobs - just install and configure clang etc. as needed
* enabled LLD for linking

ci-exactly: build_test_runtime