@ScottTodd 2024-08-01
Timing (low sample size):

Platform | Baseline time | Time with LLD | Time with LLD + ccache
-- | -- | -- | --
ubuntu | 9m45s | 4m30s | 1m30s
windows | 27m | 8m30s | 3m
macos | 5m30s | 4m15s | 1m30s


Baseline: https://github.com/iree-org/iree/actions/runs/10203742888/
LLD: https://github.com/iree-org/iree/actions/runs/10206978325/
LLD + ccache: https://github.com/ScottTodd/iree/actions/runs/10206588680/